# swagger.yml
openapi: 3.1.0
info:
  title: Receipt Wrangler API.
  termsOfService:
    there are no TOS at this moment, use at your own risk we take no
    responsibility
  license:
    name: GPLV3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  version: 5.0.0
servers:
  - url: /api
security:
  - bearerAuth: [ ]
paths:
...
  /receiptImage/:
    post:
      tags:
        - ReceiptImage
      summary: Uploads a receipt image
      description: This will upload a receipt image, [SYSTEM USER]
      requestBody:
        description: Receipt image to upload
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ReceiptFileUploadCommand"
      operationId: uploadReceiptImage
      responses:
        200:
          description: The uploaded receipt image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDataView"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
  /receiptImage/convertToJpg:
    post:
      tags:
        - ReceiptImage
      summary: Converts a receipt image to jpg
      description: This will convert a receipt image to jpg, [SYSTEM USER]
      requestBody:
        description: Receipt image to convert
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Base64 encoded image
      operationId: convertToJpg
      responses:
        200:
          description: The converted image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EncodedImage"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
...
  /receiptImage/{receiptImageId}:
    parameters:
      - in: path
        name: receiptImageId
        schema:
          type: integer
        required: true
        description: Id of receipt image to get
    get:
      tags:
        - ReceiptImage
      summary: Get receipt image
      description: This will get a receipt image by id, [SYSTEM USER]
      operationId: getReceiptImageById
      responses:
        200:
          description: The receipt image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileDataView"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
    delete:
      tags:
        - ReceiptImage
      summary: Delete receipt image
      description: This will delete a receipt image by id [SYSTEM USER]
      operationId: deleteReceiptImageById
      responses:
        200:
          $ref: "#/components/responses/Ok"
        500:
          $ref: "#/components/responses/Internal"
      security:
        - bearerAuth: [ ]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Request was successfully processed
...
    Internal:
      description: There was an error processing the request
  schemas:
...
    BaseModel:
      type: object
      required:
        - id
        - createdAt
      properties:
        id:
          type: integer
          format: uint64
        createdAt:
          type: string
        createdBy:
          type: integer
          format: uint64
          default: 0
        createdByString:
          type: string
          description: Created by entity's name
          default: ""
        updatedAt:
          type: string
          default: ""
...
    ReceiptFileUploadCommand:
      type: object
      required:
        - file
        - receiptId
      properties:
        file:
          type: string
          format: binary
        receiptId:
          type: integer
          description: Receipt foreign key
        encodedImage:
          type: string
          description: Base64 encoded image for file types that aren't viewable natively in the browser, such as PDFs
    FileDataView:
      allOf:
        - $ref: "#/components/schemas/BaseModel"
        - type: object
          required:
            - encodedImage
            - name
          properties:
            encodedImage:
              type: string
              description: Base64 encoded image
            name:
              type: string
              description: File name
    EncodedImage:
      type: object
      required:
        - encodedImage
      properties:
        encodedImage:
          type: string
          description: base64 encoded jpg
...